backend/requirements.txt
flask
flask-cors
google-generativeai
pydantic
python-dotenv
requests
dotenv
google-genai
flask-sqlalchemy